add parser.c
