+++ Makefile
@@ -6,6 +6,7 @@ OBJ_PATH	= obj/
 SRC_EXEC_PATH = srcs/exec/
 SRC_BUILTINS_PATH = srcs/builtins/
 SRC_LEXER_PATH = srcs/lexer/
+SRC_PARSER_PATH = srcs/parser/
 
 CC			= cc
 CFLAGS		= -Wall -Werror -Wextra -MMD -MP -g3
@@ -21,16 +22,19 @@ SRC_BUILTINS = echo.c \
 				exit.c
 SRC_LEXER = lexer.c \
 			lst_lexer.c
+SRC_PARSER = parser.c
 
 SRCS	  = $(addprefix $(SRC_PATH), $(SRC))
 SRCS_EXEC = $(addprefix $(SRC_EXEC_PATH), $(SRC_EXEC))
 SRCS_BUILTINS = $(addprefix $(SRC_BUILTINS_PATH), $(SRC_BUILTINS))
 SRCS_LEXER = $(addprefix $(SRC_LEXER_PATH), $(SRC_LEXER))
+SRCS_PARSER =$(addprefix $(SRC_PARSER_PATH), $(SRC_PARSER))
 OBJ		  = $(SRC:.c=.o)
 OBJ_EXEC  = $(SRC_EXEC:.c=.o)
 OBJ_BUILTINS = $(SRC_BUILTINS:.c=.o)
 OBJ_LEXER = $(SRC_LEXER:.c=.o)
-OBJS	  = $(addprefix $(OBJ_PATH), $(OBJ) $(OBJ_EXEC) $(OBJ_BUILTINS) $(OBJ_LEXER))
+OBJ_PARSER = $(SRC_PARSER:.c=.o)
+OBJS	  = $(addprefix $(OBJ_PATH), $(OBJ) $(OBJ_EXEC) $(OBJ_BUILTINS) $(OBJ_LEXER) $(OBJ_PARSER))
 DEPS	  = $(addprefix ${OBJ_PATH}, ${SRC:.c=.d})
 HEADERS	  = -I./headers/
 LIB 	  = -L./libft_42/ -lft -lreadline
@@ -59,6 +63,10 @@ $(OBJ_PATH)%.o: $(SRC_BUILTINS_PATH)%.c
 	@$(CC) $(HEADERS) -o $@ -c $<
 
 $(OBJ_PATH)%.o: $(SRC_LEXER_PATH)%.c
+	@mkdir -p $(OBJ_PATH)
+	@$(CC) $(HEADERS) -o $@ -c $<
+
+$(OBJ_PATH)%.o: $(SRC_PARSER_PATH)%.c
 	@mkdir -p $(OBJ_PATH)
 	@$(CC) $(HEADERS) -o $@ -c $<
