+++ headers/exec.h
@@ -1,8 +1,9 @@
 #ifndef EXEC_H
 # define EXEC_H
 # include <stdio.h>
-# include "../libft_42/includes_libft/libft.h"
-# include "lexer.h"
+# include "main.h"
+//# include "../libft_42/includes_libft/libft.h"
+//# include "lexer.h"
 # include "builtins.h"
 
 void	read_input(t_data *data);
