+++ headers/minishell.h
@@ -1,23 +1,28 @@
 #ifndef MINISHELL_H
 # define MINISHELL_H
-# include <stdio.h>
-# include <stdlib.h>
-# include <readline/readline.h>
-# include <readline/history.h>
+//# include <stdio.h>
+//# include <stdlib.h>
+//# include <readline/readline.h>
+//# include <readline/history.h>
+# include "main.h"
+//# include "../libft_42/includes_libft/libft.h"
+//# include "../libft_42/includes_libft/garbage.h"
 # include "exec.h"
 # include "lexer.h"
+# include "parser.h"
 
 //t_data g_data;
 
-# ifndef T_DATA
-#  define T_DATA
-
-typedef struct s_data
-{
-	char *input;
-	t_list *garb_lst;
-	t_token *lexer_lst;
-}	t_data;
-# endif
+//# ifndef T_DATA
+//#  define T_DATA
+//
+//typedef struct s_data
+//{
+//	char *input;
+//	t_list *garb_lst;
+//	t_token *lexer_lst;
+//	t_elem_pars	*parser_lst;
+//}	t_data;
+//# endif
 
 #endif
