+++ srcs/parser/parser.c
@@ -1,0 +1,8 @@
+#include "../../headers/parser.h"
+
+extern t_data	g_data;
+
+void	parser(t_data *data)
+{
+
+}
